Move BBVAS eBill modules into eBill namespace
Rename and relocate eBill-related modules from src/BBVAS/* to src/eBill/* (BillStatusRequest, SMSServiceStatusRequest, SmartBillRegistration and SmartBillRegistration Source — controllers, routes, services). Update src/app.js to remove BBVAS imports and require the new eBill paths, and rewire route mounts (smartBillRegistration, bill status and SMS status endpoints) to the eBill modules. This organizes eBill functionality under the eBill folder and updates application wiring accordingly.
